Expose some helpers for external cache management
- Request.noCacheReponse()
    - Bypass the cache for the current request.
- Request.getCacheEntry()
    - Get the CacheEntry for the current request.
